+++ sdk/cosmos/azure-cosmos/src/test/java/com/azure/cosmos/CosmosDiagnosticsTest.java
@@ -5,6 +5,7 @@ package com.azure.cosmos;
 
 import com.azure.core.http.ProxyOptions;
 import com.azure.cosmos.implementation.AsyncDocumentClient;
+import com.azure.cosmos.implementation.ClientSideRequestStatistics;
 import com.azure.cosmos.implementation.Configs;
 import com.azure.cosmos.implementation.DatabaseForTest;
 import com.azure.cosmos.implementation.HttpConstants;
@@ -299,12 +300,15 @@ public class CosmosDiagnosticsTest extends TestSuiteBase {
 
     @Test(groups = {"simple"}, dataProvider = "query", timeOut = TIMEOUT)
     public void queryMetrics(String query, Boolean qmEnabled) {
+        CosmosContainer directContainer =
+            this.directClient.getDatabase(this.cosmosAsyncContainer.getDatabase().getId())
+                .getContainer(this.cosmosAsyncContainer.getId());
         CosmosQueryRequestOptions options = new CosmosQueryRequestOptions();
         if (qmEnabled != null) {
             options.setQueryMetricsEnabled(qmEnabled);
         }
         boolean qroupByFirstResponse = true; // TODO https://github.com/Azure/azure-sdk-for-java/issues/14142
-        Iterator<FeedResponse<InternalObjectNode>> iterator = this.container.queryItems(query, options,
+        Iterator<FeedResponse<InternalObjectNode>> iterator = directContainer.queryItems(query, options,
             InternalObjectNode.class).iterableByPage().iterator();
         assertThat(iterator.hasNext()).isTrue();
         while (iterator.hasNext()) {
@@ -313,7 +317,68 @@ public class CosmosDiagnosticsTest extends TestSuiteBase {
             assertThat(feedResponse.getResults().size()).isEqualTo(0);
             if (!query.contains("group by") || qroupByFirstResponse) { // TODO https://github
                 validateQueryDiagnostics(queryDiagnostics, qmEnabled, true);
+                validateDirectModeQueryDiagnostics(queryDiagnostics);
+                if (query.contains("group by")) {
+                    qroupByFirstResponse = false;
+                }
+            }
+        }
+    }
 
+    private void validateDirectModeQueryDiagnostics(String diagnostics) {
+        assertThat(diagnostics).contains("\"connectionMode\":\"DIRECT\"");
+        assertThat(diagnostics).contains("supplementalResponseStatisticsList");
+        assertThat(diagnostics).contains("responseStatisticsList");
+        assertThat(diagnostics).contains("\"gatewayStatistics\":null");
+        assertThat(diagnostics).contains("addressResolutionStatistics");
+        assertThat(diagnostics).contains("\"userAgent\":\"" + Utils.getUserAgent() + "\"");
+    }
+
+    private void validateGatewayModeQueryDiagnostics(String diagnostics) {
+        assertThat(diagnostics).contains("\"connectionMode\":\"GATEWAY\"");
+        assertThat(diagnostics).doesNotContain(("\"gatewayStatistics\":null"));
+        assertThat(diagnostics).contains("\"operationType\":\"Query\"");
+        assertThat(diagnostics).contains("\"userAgent\":\"" + Utils.getUserAgent() + "\"");
+        assertThat(diagnostics).contains("\"regionsContacted\"");
+    }
+
+    @Test(groups = {"simple"}, dataProvider = "query", timeOut = TIMEOUT*2)
+    public void queryDiagnosticsGatewayMode(String query, Boolean qmEnabled) {
+        CosmosClient testDirectClient = new CosmosClientBuilder()
+                                            .endpoint(TestConfigurations.HOST)
+                                            .key(TestConfigurations.MASTER_KEY)
+                                            .contentResponseOnWriteEnabled(true)
+                                            .gatewayMode()
+                                            .buildClient();
+        CosmosQueryRequestOptions options = new CosmosQueryRequestOptions();
+
+        CosmosContainer cosmosContainer = testDirectClient.getDatabase(cosmosAsyncContainer.getDatabase().getId())
+                                              .getContainer(cosmosAsyncContainer.getId());
+        List<String> itemIdList = new ArrayList<>();
+        for (int i = 0; i < 100; i++) {
+            InternalObjectNode internalObjectNode = getInternalObjectNode();
+            CosmosItemResponse<InternalObjectNode> createResponse = cosmosContainer.createItem(internalObjectNode);
+            if (i % 20 == 0) {
+                itemIdList.add(internalObjectNode.getId());
+            }
+        }
+        boolean qroupByFirstResponse = true;
+        if (qmEnabled != null) {
+            options.setQueryMetricsEnabled(qmEnabled);
+        }
+        Iterator<FeedResponse<InternalObjectNode>> iterator = cosmosContainer
+                                                                  .queryItems(query, options, InternalObjectNode.class)
+                                                                  .iterableByPage()
+                                                                  .iterator();
+        assertThat(iterator.hasNext()).isTrue();
+
+        while (iterator.hasNext()) {
+            FeedResponse<InternalObjectNode> feedResponse = iterator.next();
+            String queryDiagnostics = feedResponse.getCosmosDiagnostics().toString();
+            assertThat(feedResponse.getResults().size()).isEqualTo(0);
+            if (!query.contains("group by") || qroupByFirstResponse) { // TODO https://github
+                validateQueryDiagnostics(queryDiagnostics, qmEnabled, true);
+                validateGatewayModeQueryDiagnostics(queryDiagnostics);
                 if (query.contains("group by")) {
                     qroupByFirstResponse = false;
                 }
@@ -344,7 +409,7 @@ public class CosmosDiagnosticsTest extends TestSuiteBase {
         String queryDiagnostics,
         Boolean qmEnabled,
         boolean expectQueryPlanDiagnostics) {
-        if (qmEnabled == null || qmEnabled == true) {
+        if (qmEnabled == null || qmEnabled) {
             assertThat(queryDiagnostics).contains("Retrieved Document Count");
             assertThat(queryDiagnostics).contains("Query Preparation Times");
             assertThat(queryDiagnostics).contains("Runtime Execution Times");